integrating requested changes
