+++ plugin/accel_embree/accel_embree.cpp
@@ -7,7 +7,7 @@
 #include <lm/scene.h>
 #include <lm/mesh.h>
 #include <lm/exception.h>
-#include "embree.h"
+#include "embree_params.h"
 
 LM_NAMESPACE_BEGIN(LM_NAMESPACE)
 
@@ -40,19 +40,16 @@ private:
 public:
 
      virtual void construct(const Json& prop) override {
-        settings_ = rtcDefaultBuildArguments();
-        from_json(prop,settings_);
+        settings_ = prop;
+        sf_ = prop;
 
-        from_json(prop,sf_);
-
+        //check actual values used for building
         Json j;
-        to_json(j,settings_);
+        j = settings_;
         LM_INFO(j.dump());
         j.clear();
-        
-        to_json(j,sf_);
+        j = sf_;
         LM_INFO( j.dump() );
-        
         }
 
     Accel_Embree() {
@@ -88,6 +85,7 @@ public:
 
         rtcSetSceneFlags(scene_, sf_);
         rtcSetSceneBuildQuality(scene_, settings_.buildQuality);
+        rtc_AT_SetNextBVHArguments(scene_, settings_);
 
         // Flatten the scene graph and setup geometries
         LM_INFO("Flattening scene");
@@ -102,7 +100,6 @@ public:
             // Record flattened primitive
             const int flatten_node_index = int(flattened_nodes_.size());
             flattened_nodes_.push_back({ Transform(global_transform), node.index });
-
             // Create triangle mesh
             auto geom = rtcNewGeometry(device_, RTC_GEOMETRY_TYPE_TRIANGLE);
             const int num_triangles = node.primitive.mesh->num_triangles();
