+++ plugin/accel_embree/accel_embree_instanced.cpp
@@ -7,7 +7,7 @@
 #include <lm/scene.h>
 #include <lm/mesh.h>
 #include <lm/exception.h>
-#include "embree.h"
+#include "embree_params.h"
 
 LM_NAMESPACE_BEGIN(LM_NAMESPACE)
 
@@ -42,18 +42,16 @@ private:
 
 public:
     
-     virtual void construct(const Json& prop) override {
-        settings_ = rtcDefaultBuildArguments();
-        from_json(prop,settings_);
-
-        from_json(prop,sf_);
-
+     virtual void construct(const Json& prop) override {        
+        settings_ = prop;
+        sf_ = prop;
+        
+        //check actual values used for building
         Json j;
-        to_json(j,settings_);
+        j = settings_;    
         LM_DEBUG(j.dump());
         j.clear();
-        
-        to_json(j,sf_);
+        j = sf_;
         LM_DEBUG(j.dump() );
         
         }
@@ -180,7 +178,8 @@ public:
             rtcscene = rtcNewScene(device_);
             rtcSetSceneFlags(rtcscene, sf_);
             rtcSetSceneBuildQuality(rtcscene, settings_.buildQuality);
-
+            rtc_AT_SetNextBVHArguments(scene_, settings_);
+    
             // Create triangle meshes
             for (const auto& fnode : fscene) {
                 // Primitive
